move configuration files for performance, permissions, static assets, tasks, and UI settings
- Introduced `misc.py` for performance and CORS configurations, including thread pool size and CORS origin validation.
- Created `permissions.py` to manage user permissions for workspaces, public sharing, chat controls, and features.
- Added `static.py` to handle static file management, including copying frontend assets and legacy support for custom names.
- Implemented `tasks.py` for task-related configurations, including title, tags, image prompt generation, and autocomplete settings.
- Established `ui.py` for UI configurations, including default locale, user roles, and prompt suggestions.
- Updated `main.py` to include the new WEBUI_AUTH configuration.
