+++ backend/open_webui/config/__init__.py
@@ -7,9 +7,15 @@ from .auth import *
 from .media import *
 from .retrieval import *
 from .vector import *
-from .legacy import *
 from .code_interpreter import *
 from .storage import *
+from .admin import *
+from .ui import *
+from .tasks import *
+from .static import *
+from .permissions import *
+from .misc import *
+from .features import *
 
 setup_logging()
